Extract environment/config parsing into config.py
Moves the read-only configuration out of main.py into a dedicated config
module: version, GitHub update-checker settings, UI theme, ALWAYS_NOTIFY_OPEN,
the multi-line .env parser, TestFlight constants, the parsed ID_LIST /
APPRISE_URLS, SLEEP_TIME, TITLE_REGEX, and HEARTBEAT_INTERVAL. config.py owns
load_dotenv() now. main.py imports these and keeps all mutable runtime state
(live ID/URL lists, Apprise object, HTTP session, metrics).

WEB_USERNAME/WEB_PASSWORD intentionally stay in main.py: they're read with
os.getenv at module level so that reloading main (as the auth test does) picks
up changed env, which wouldn't happen via the import-cached config module.

Behaviour unchanged; main.py trimmed by ~90 more lines (~1700 total).
